Fix airflow-ctl workspace and include devel-common again along with pytest_plugins to make breeze testing work
