+++ airflow-ctl/pyproject.toml
@@ -93,6 +93,13 @@ exclude_also = [
 codegen = [
     "datamodel-code-generator[http]==0.28.2",
 ]
+dev = [
+    "apache-airflow-devel-common",
+]
+[tool.uv.sources]
+# These names must match the names as defined in the pyproject.toml of the workspace items,
+# *not* the workspace folder paths
+apache-airflow-devel-common = {workspace = true}
 
 # To use:
 #
