+++ airflow-ctl/tests/airflow_ctl/commands/test_auth_command.py
@@ -21,7 +21,6 @@ import os
 from unittest import mock
 from unittest.mock import patch
 
-import pytest
 from platformdirs import user_config_path
 
 from airflow.ctl.cli import cli_parser
@@ -29,7 +28,6 @@ from airflow.ctl.cli.commands import auth_command
 
 
 # TODO: Remove this flag when all CLI integrated with API.
-@pytest.mark.db_test
 class TestCliAuthCommands:
     parser = cli_parser.get_parser()
